requirements.txt
django==3.2
django-imagekit==4.0.2
setuptools
legacy-cgi
python-dotenv
